+++ app/page.tsx
@@ -3,7 +3,7 @@
 import { useState } from "react";
 import CurrentWeather from "./components/CurrentWeather";
 import Search from "./components/Search";
-import { IMAGE_URL, WEATHER_API_URL } from "./config";
+import { IMAGE_URL, WEATHER_API_URL } from "./utils/config";
 import { SearchData } from "./models/apiTypes";
 import { CurrentWeatherData } from "./models/componentProps";
